[FEAT/#228] 페어프로그래밍 Follow 관련 최종
